Created Functions for Actions
More modular. Separated functions that can be called for forward, stop, turn left, turn right. Did not alter setup().  Need to define delay() for how long it takes the robot to make a 90 degree turn. I'm pretty sure this is the right set up for calling functions and etc in arduino.